+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - Hook for underlying statistics emitted from librdkafka have been added.
 - Hook for underlying async errors emitted from  librdkafka have been added.
 - ActiveJob Rails adapter.
+- Added benchmarks that can be used to profile Karafka.
 
 ## 1.4.11 (2021-12-04)
 - Source code metadata url added to the gemspec